requirements.txt
streamlit
pandas
requests
reportlab
streamlit-authenticator
google-genai